A bunch of changes to manage saving checkpoint files in MapFragment.
There are a number of structural changes to MapFragment and you will need to clear it out before you try to generate using this version.

testHarness?action=testAPKml 

And loadLots will reload the proper tables.

This version isn't async yet.  This is just for testing the checkpoints and all the zip files.