Issue #280 - Reuse the version name from AndroidManifest.xml
Get rid of the version name in res/strings.xml in order to reuse the
same attribute from AndroidManifest.xml. When checking for new versions
or displaying app info to the user, the version name will be the one in
the manifest, allowing us to maintain a single value across the app.