[#320] NO_IMEI should be treated as null
On WiFi only mobile devices (no SIM, no imei) the dashboard was storing
"NO_IMEI" as value, and that makes all this devices to be handled as the
same.

If we try to get a device by `imei` and is "NO_IMEI" we immediatly
return `null`, and the related code needs to search by `phoneNumber`.

`phoneNumber` property stores either the actual mobile phone number or
the MAC address of the device.