[#650] Strategy for validating raw data imports
Loop through the header cells until a cell matches "[0-9]+\\|.+". This
is the first question header cell and must be at index 4
(non-monitoring) or index 6 (monitoring). If no such question is found
validation fails. The rest of the header cells must match the pattern
".+\\|.+". We break the loop on the first validation error.