[#2558] Deleting deprecated SurveyBulkUploader
To produce the jar, we want to remove the maven-compiler exclusion, so
that we can use the maven-war plugin to package all the classes

SurveyBulkUploader depends on the ProgressDialog that depends on AWT,
and we do not want this dependency on the Flow war.

SurveyBulkUploader has been deprecated since 2014 and is not used
anywhere.