[#2628] Updating commands for restarting the dev server
Using appengine:stop and appengine:run.

Note that while in the devserver.sh we use appengine:start, so that the server
is started in the background, when running the docker-compose exec command
we are using appengine:run. This is because docker-compose exec when it
finish running all the commands, kills the session, which includes the
dev server started in the background. "nohup" doesnt seem to work.

appengine:run do not exit, so the docker-compose command now specifies a `-d`
for the command to run in detach mode.

As we running in detach mode doesnt show the stdout of the command, we now
redirect the output to a build.log file so it shows up in the docker-compose logs

Also updating all docker-compose commands to be run as akvo user

If not, they will run as root which will cause both some issues with
permissions and for the commands to with the home directory "/root" which
is not a mounted volume, which means that the ~/.m2 directory is not
shared with the host machine, which means that the maven deps are
downloaded every time the container is recreated