[#2403] Set the current-user-id with the id returned by the backend
After clicking "save" new user, stay on the same edit form
and once the backend returns the user-id, set it as the
current-user-id (the id of the user being edited).

Added a random value to the list-user state to avoid a possible race
condition if the partner clicks "back" faster than the backend takes
to return the newly created user.

There is still a race condition if the partner creates a new user, clicks back,
clicks create new user and fills up some of the form before the first
new user backend call finished. In this case, the new user form is cleaned. Fixing
this race condition requires a major change in the code.