[#2823] Remove the instanceCount from response
With a large enough number of forms (aka Survey) we go and get the
number of counts for each one of them reaching the limit of 60s to
respond in GAE.

This is just the first patch to fix the symptom.

It seems that we're not using at all the SurveyInstanceSummary, so all
the code relatated to it should be removed.