[#3031] Use current time for events
Instead of using the last updated time of the Entity as the event timestamp,
just use the current time, that is, the time when this script is run.

This is so that if Unilog has already processed some events after the Entity was updated, this new events (which are old entities) are still processed.

As an alternative, we could "reset" the Unilog offsets, but this will mean rereading all the
EventQueue data, which is deemed as expensive.