[#3496] Adds countBySurveyGroupId
Unfortunately we can't use JDO with `KeysOnly` queries.

The current documentation states the following:

    Query q = new Query("Person").setKeysOnly();

Translates to this query in JDO

    Query q = pm.newQuery("select id from " + Person.class.getName());
    List<String> ids = (List<String>) q.execute();

When trying this approach DataNucleus fails with

   "Variable expression is not supported by this mapper"

Since we're using the "low-level" API at the BaseDAO class,
we implemented a generic countByFilterProperty as a way of doing
an efficient Key-Only query.