[#1223] Use a 'transaction' for new Assignment
The new `Assignment` behavior is different than the rest of the
application, e.g.  When creating a new `Survey`, we create and commit
the record, we get a the new `keyId` from the server and then we ask
the user to "edit" it.

In the case of `Assignment` we create a new record in the client and
we don't commit it until the user clicks the "Save" button or the user
navigates away triggering an implicit `FLOW.store.commit()`.

Since we don't want to save partially filled `Assignment` if the user
navigates away, we want to discard that record.

Ember Data has a concept of "transaction" and we use it to deal with
this exception.  When creating a new `Assignment` we do it within a
transaction and we discard if the user navigates away.

>  Because newly created records are dirty from the time they are created,
>  and because dirty records can not be added to a transaction, you must
>  use the `createRecord()` method to assign new records to a transaction.
>  For example, instead of this:
>    var transaction = store.transaction();
>    var person = App.Person.createRecord({ name: "Steve" });
>    // won't work because person is dirty
>    transaction.add(person);
>  Call `createRecord()` on the transaction directly:
>    var transaction = store.transaction();
>    transaction.createRecord(App.Person, { name: "Steve" });

https://github.com/akvo/akvo-flow/blob/20200403-170810/Dashboard/app/js/vendor/ember-data-rev10.js#L516-L555