[#3524] Use `equals` instead of `==` for String comparison
A common pitfall when trying to compare strings is to use `==`
operator, but this one is for checking if to objects are the same
"reference".  Most of the cases we want to use `.equals()` function.

In this particular issue the check `!= ""` was evaluating that the
passed string parameter `since` was the same refence, so the code
prepared a cursor of only 20 entities.

Fixing this problem now leads to a different UX problem that the user
needs to click one by one a pontential large list of checkboxes. This
problem will be fixed in a different changeset.

More info: https://stackoverflow.com/questions/513832/how-do-i-compare-strings-in-java/513839#513839