[#3479] Specify `cloudSdkHome` property for Maven plugin
As per documentation, by specifying the `cloudSdkHome` we avoid the
plugin to go and download the Google SDK everytime. We already have
the SDK installed as part of the container.

This will speed up build + deployments since `mvn package` and `mvn
appengine:stage` will assume that there is a Google SDK at
`/google-cloud-sdk`

https://github.com/GoogleCloudPlatform/app-maven-plugin/blob/d6d0638ae7bf454eb6a30177c12367132d6a4cfd/USER_GUIDE.md#cloud-sdk-configuration