[#3571] Use "flip-" and "promote-" tags
We keep the same flow/logic as before but we are going to change the tag names to match their meaning with what we do in other projects.

1. "promote-" tag will mean to deploy to dark production. It additionally will flip UAT1 to make it easy to access the new version, as right now to access a dark instance you need to guess a very big and per-version-specific url
1. "flip-" tags will flip the traffic to the new version.

Moving all the logic regarding which CI steps to run to Semaphore. The trade off is easier to understand what is CI running + no duplication vs no local testing + less portable scripts.